iago/requirements.txt
django==3.2.8
djangorestframework==3.12.4
psycopg2-binary==2.8.6
daphne==3.0.1
whitenoise==5.3.0
jsonschema
requests==2.26.0
twisted[tls,http2]
langdetect==1.0.9
scrapy==2.5.1
beautifulsoup4==4.10.0
bertopic==0.9.3
boto3==1.20.7
faiss-cpu==1.7.1.post2 
drf-yasg==1.20.0 
git+https://github.com/scikit-learn-contrib/hdbscan.git@1c24b8ddb787f6878ec797f5d14654e041dd352f